Added base for shop
* Added shop entity
* Press E to open shop while aiming at entity
* Open shop menu
* Click buy item to buy item
* Removes X amount of money and gives item to inventory

+++ Assets/ModularFirstPersonController/FirstPersonController/PlayerStats.cs
@@ -4,7 +4,7 @@ using TMPro;
 public class PlayerStats : MonoBehaviour
 {
     // Money variables
-    public int money = 1000;
+    public float money = 1000;
     public TMP_Text moneyTextMesh;
 
     // Health variables
